reorg(4/5): rename services/ → api/ + classic.ts/home.ts
Renames the services/ folder to api/, which is both more accurate
(the folder contains literally the two MELCloud APIs the package
wraps) and aligned with the package name `@olivierzal/melcloud-api`.
Also drops the historical asymmetry of `api.ts` vs `home-api.ts` in
favor of a symmetric `classic.ts` / `home.ts` pair named after the
two MELCloud backends:

- `services/api.ts`          → `api/classic.ts`
- `services/home-api.ts`     → `api/home.ts`
- `services/interfaces.ts`   → `api/interfaces.ts`
- `services/sync-manager.ts` → `api/sync-manager.ts`
- `services/index.ts`        → `api/index.ts`

The exported class names `MELCloudAPI` and `MELCloudHomeAPI` are
unchanged — only the files move and their paths update. The public
API surface (main.ts re-exports) remains stable.

`services/` is fully removed after this commit.
