+++ src/facades/factory.ts
@@ -1,5 +1,5 @@
+import type { APIAdapter } from '../api/index.ts'
 import type { DeviceModelAny, Model, ModelRegistry } from '../models/index.ts'
-import type { APIAdapter } from '../services/index.ts'
 import { DeviceType } from '../constants.ts'
 import type { DeviceFacadeAny, Facade } from './interfaces.ts'
 import { AreaFacade } from './area.ts'
