+++ src/main.ts
@@ -102,6 +102,30 @@ export {
   verticalToClassic,
 } from './adapters/index.ts'
 export {
+  type API,
+  type APIAdapter,
+  type APIConfig,
+  type APISettings,
+  type BaseAPIConfig,
+  type ErrorDetails,
+  type ErrorLog,
+  type ErrorLogQuery,
+  type HomeAPI,
+  type HomeAPIConfig,
+  type HomeAPISettings,
+  type Logger,
+  type OnSyncFunction,
+  type RequestCompleteEvent,
+  type RequestErrorEvent,
+  type RequestLifecycleContext,
+  type RequestLifecycleEvents,
+  type RequestRetryEvent,
+  type RequestStartEvent,
+  type SettingManager,
+  MELCloudAPI,
+  MELCloudHomeAPI,
+} from './api/index.ts'
+export {
   type NonSilentFanSpeed,
   coolModes,
   DeviceType,
@@ -169,27 +193,3 @@ export {
   isDeviceOfType,
   ModelRegistry,
 } from './models/index.ts'
-export {
-  type API,
-  type APIAdapter,
-  type APIConfig,
-  type APISettings,
-  type BaseAPIConfig,
-  type ErrorDetails,
-  type ErrorLog,
-  type ErrorLogQuery,
-  type HomeAPI,
-  type HomeAPIConfig,
-  type HomeAPISettings,
-  type Logger,
-  type OnSyncFunction,
-  type RequestCompleteEvent,
-  type RequestErrorEvent,
-  type RequestLifecycleContext,
-  type RequestLifecycleEvents,
-  type RequestRetryEvent,
-  type RequestStartEvent,
-  type SettingManager,
-  MELCloudAPI,
-  MELCloudHomeAPI,
-} from './services/index.ts'
