+++ tests/unit/home-device-ata-facade.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest'
 
-import type { HomeAPI } from '../../src/services/interfaces.ts'
+import type { HomeAPI } from '../../src/api/interfaces.ts'
 import type { HomeDevice } from '../../src/types/index.ts'
 import { HomeDeviceType } from '../../src/constants.ts'
 import { HomeDeviceAtaFacade } from '../../src/facades/home-device-ata.ts'
